+++ cpp/inc/bond/core/bonded.h
@@ -30,7 +30,7 @@ ApplyTransform(const Transform& transform, const bonded<T, Reader>& bonded);
 // Helper function move_data for dealing with [not] moving a Reader& in bonded<T, Reader&>
 template <typename T, typename U>
 typename boost::enable_if<std::is_reference<T>, T>::type
-inline move_data(U& data)
+inline move_data(U& data) BOND_NOEXCEPT
 {
     BOOST_STATIC_ASSERT(std::is_same<T, U&>::value);
     return data;
@@ -38,7 +38,8 @@ inline move_data(U& data)
 
 template <typename T, typename U>
 typename boost::disable_if<std::is_reference<T>, T&&>::type
-inline move_data(U& data)
+inline move_data(U& data) BOND_NOEXCEPT_IF(
+    std::is_nothrow_move_constructible<T>::value)
 {
     BOOST_STATIC_ASSERT(std::is_same<T, U>::value);
     return std::move(data);
@@ -77,7 +78,7 @@ public:
 
     /// @brief Move constructor
     bonded(bonded&& bonded) BOND_NOEXCEPT_IF(
-        std::is_nothrow_move_constructible<Reader>::value
+        BOND_NOEXCEPT_EXPR(detail::move_data<Reader>(bonded._data))
         && std::is_nothrow_move_constructible<RuntimeSchema>::value)
         : _data(detail::move_data<Reader>(bonded._data)),
           _schema(std::move(bonded._schema)),