+++ cpp/inc/bond/core/bonded_void.h
@@ -50,7 +50,7 @@ public:
 
     /// @brief Move constructor
     bonded(bonded&& other) BOND_NOEXCEPT_IF(
-        std::is_nothrow_move_constructible<Reader>::value
+        BOND_NOEXCEPT_EXPR(detail::move_data<Reader>(other._data))
         && std::is_nothrow_move_constructible<RuntimeSchema>::value)
         : _data(detail::move_data<Reader>(other._data)),
           _schema(std::move(other._schema)),